Add ConcurrentHashMapOfLists and use it for Combat.attackedEntities
Fixes bug 740, intermittent ConcurrentModificationException

The alternative would be synchronizing all access to attackedEntities,
either at the method level or by using an internally synchronized
container like a Hashtable.  But ConcurrentHashMap should give better
performance, since it only locks on writes not reads, and we read
attackedEntities a lot more often than we write to it.