Switch game thread executor to be a single-thread executor instead, of a thread pool. This is needed because when things need to execute on the game thread, it's not thread safe so only one such thing should be executing at a time.
This should fix ConcurrentModificationException
from CardAreaPanel.selectCard() on mobile.

This is a reland of r34680 with a fix for being able
to run tasks on the game thread while waiting for input
from the user. Specifically, tested paying costs and
using the dev panel, but should work with anything that
schedules tasks via GameAction.invoke().