CHANGES TO KEYWORDS !!!DANGER!!!
I'm messing with powers I don't understand. Here I go.

Following changes to menace that removed reminder text by removing the keyword.equals("Menace"), I conclude that you can add reminder text by inserting the keyword.equals("Keyword"). This is part one if it works, as it is for all the simple keywords that don't rely on numbers or costs (Compare Devoid with Awaken)

If this doesn't work, and everything breaks, burn it.