Added Experience Counters to Desktop Forge
Experience counters are at the lowest priority from Poison>Energy>Experience. It uses the Commander 2015 symbol, in which Experience Counters appeared.
There was a lot of transplanting involved copying the Energy code to fit the Experience model. This means there might be some unforseen problems though so check it out. I did several run throughs with Aetherwind Basker, Attune with Aether, Caress of Phyrexia, Leeches, and the Experience counter commanders to find anything remaining. There's a difference between "There are no bugs" and "I found no bugs".

Speaking of Leeches, there was an odd bug in the ProcessRunLogThing, saying "SVar 'X' not found in ability, fallback to Card (Leeches). Ability is ()". Something to do with Svar:X:TargetedPlayer$PoisonCounters maybe?