Improve reported multiplayer host IP
The existing IP selection logic was less than optimal on Linux hosts
where it would frequently select and report 127.0.0.1 as the IP address
to share.  The new logic will create a datagram socket, use it to locate
an interface with a default route, and further backtrack to the
interface's IPv4 address.

Signed-off-by: Jamin W. Collins <jamin.collins@gmail.com>