Do not use Fireslinger on players if our life lower Do prefer targeting the player if has a "when this deals damage to an opponent" trigger
Important : The AI should be able to recognize the ability stacks - if there are multiple cards that deal damage, they should be used to kill the target together. (for example, tap two prodigal sorcerers to kill a 2/2 creature!)
Unfortunately, I don't think I can implement this with my current knowledge.