This resolves an issue when hashing NON-existing cards.
Hash code now considers collectorNumber, which is retrieved from card edition, if not available. For non existing cards, the edition won't be found (null return) which resulted in a NPE.

This patch addresses this situation.
In future, this method will be removed, and PaperCard created directly with corresponding collectorNumber from CardInSet.