New Card Types categories also for custom editions
Card Editions now includes updated Types, following the new remapping.

Moreover, a new option to CardEdition.Reader has been added (in the constructor) in order to account for custom editions.

So far this new "reading mode" accounts for type that is "forced" always to be "THIRDPARTY" so to guarantee consistent sets filtering (so far only on Mobile platform).
This new option is integrated in StaticData when instantiating the collection for CustomEditions.