Brand new desktop UI Component representing a JTree of FCheckboxes
Differently from standard JTree, this component leverages on a custom TreeCellRenderer that renders each node with a checkbox, therefore rewriting completely the selection mechanism of the JTree.

This component is currently used in the DialogChooseSets to represent card editions grouped per edition type.
In general, this component can be used elsewhere, whenever nested objects for selection should be used.

The class exposes a public API for easy interaction, as well as customised classes for nodes and nodes data, and specialised events and listeners.