New constructor with limitedSets list for dual of unselectable
There are cases in which the set filter should be initialised with the white list of allowed set codes, rather than the blacklist (i.e. unselectableSets).

This is the case of CardCatalog or DeckCatalog with constraints or limitations on card pool (i.e. restricted cards, and or restricted sets).

This new set of constructors with extra parameters allow to handle this cases, so that the UI can easily adapt.