Unban Hasran Ogress & King Suleiman
Motivation: There are too many cards banned for AI use in the ARN set. It's something like half of the cards. This unbans 2 cards that are viable for the AI to use.

King Suleiman has no AI problems. It kills Efreets and Djinns in ARN, but also kills Changelings in newer sets. There is no reason for this card to be banned.

Hasran Ogress is banned because the AI currently will never attack using it, even when it has mana open. However, you can still do a lot worse than a 2 mana 3/2 defender. Hasran Ogress is better than several walls in Alpha, and allowing the AI to play this card will make the AI decks stronger, even when played as a defender.