Card Art optimisation algorithm for DeckImport
The card art optimisation algorithm integrated into DeckImportController leverages (in terms of general logic) on the card art optimisation in Deck.
However, several are differences in terms of context and interactions so a specialised implementation is needed.

First, the smart card art optimisation has been added as an option in DeckImport. Therefore, smart card art is not a static game option to be used when loading decks with cards with no edition, but also as an option to tweak the deck list.

The second specialisation required is that this implementation takes into account any game format constraint (in terms of allowed sets for optimisation).

Last but not least, the current deck in editor (if any) will be also considered during optimisation.
In particular, when _importing cards_ in current deck (therefore, when not creating a new deck with the deck list) will also take into account the pools of current deck so that harmonisation will already consider where cards will be imported to.