Update effluence_devourer.txt
Due to the weird wording of this card, saccing a token with the croc in play should give a trigger that does nothing. This however causes Forge to crash meaning we need a conditioncheck in line 7 to prevent Forge from giving the effect to a token while still making it trigger.