Make HarperPrism promos a single set
The HarperPrism book promo cards are part of a separate set on Scryfall.  The PHPR set code is used on numerous other sites when exporting deck lists, but this set code cannot be used in Forge deck importing because these cards were split across multiple "editions" and they can't have the same secondary code or alias.  This set organization matches the way they are categorized on Scryfall so deck imports with them in Forge will work.