Make set codes case-insensitive during deck import
This moves the createMap() function up to the StorageReaderBase class and makes all subclasses call this function in readAll() to allow subclasses to override this.  We then override this in the CardEdition Reader class to create a case-insensitive map for the card sets during deck import.  This matches the way set aliases are already using a case-insensitive map, so lowercase set codes (e.g. Moxfield and other sites exports) will work for all sets in deck import.