Speed up simulation AI code.
This change speeds up simulation AI code by as much as 30% (!) by optimizing getChangedCardTypes() when they are unchanged. That method is on the hot path of a lot of engine code, such as evaluating static and replacement effects.

On a specific game state I was testing with, it made a simulation that was taking ~26s previously to 18s after this change.

Another minor optimization is included as well, which avoids some unnecessary sb.toString() calls.