Fix Jumpstart pack based deck generation for Standard mode starts mode
Without resetting jumpStartSheetsCandidates, a mono colored deck is generated regardless of difficulty in Standard mode, and the jumpstart cards available are not reset on new game.  As a result, not only were decks being generated in unintended fashion, but if you started a new Standard game, then restarted with another color, your deck would be named as expected but generated with the previously selected deck parameters.