An update of the Constructed decks in Adventure
An update of the Constructed decks in Adventure to help with the play experience. As some of the were incredibly slow, to the point they were unplayable unless the AI got flooded or screwed. No change to number of rarities in each deck. Simply a better curve that stays in-theme to original deck theme. All have been tested in-game now to confirm they are now playable, while still being simply a starting point for a deck to come together.