Test on all current Java LTS versions (8, 11 ,17, and 21)
Tests on JDK 17 and 21 are optional (since they currently fail).

The very existence of these testes will increase the awareness on the
importance of code compatibility with newer Java versions, thus enabling
future transitions to be implemented gradually.

Making the code run on JDK 11 took a lot of hacking.
This will reduce the effort for eventual future transitions.