Update build scripts to enable multi-forge installation
Because multi-forge installation requires the app to use the generated
'R' class, if the android maven plugin isn't set up properly, building
the forge-gui-android module will fail with an import error looking for
the 'R' class.

This currently happens if the user runs any default maven target like
'test', 'verify', or 'install' without an android profile enabled. The
android profiles set up the android-maven-plugin, so running `mvn test`
for example without an android profile enabled results in the import
error.

This commit aims to correct this situation by forcing the developer to
specify that they want to build the android components by explicitly
activating the `android` profile. Then we also introduce a variable
called `android.variant` which specifies which type of apk to build. The
default is the new 'dev' build which is built from the
`android-test-build` profile. This way if the user activates the
`android` profile, at least one android apk profile will be active and
if the user doesn't have the android sdk set up properly it will yield a
meaningful error.

We also adjust application id and name of the android-test-build profile
to 'forge.app.dev' and 'Forge (dev)' so it's clear to the user which
version it is.

* Add 'android' profile to the root pom. Activating this profile adds
  the forge-gui-android module to the build and enables the
  `android-test-build` profile by default. the developer can select an
  alternate variant by specifying `-Dandroid.variant=release-build` for
  example to get the release build instead.
* Add `android.variant` variable to control which android profile is
  active.
* Adjust applicationid, application name, and sentry sources to enable
  parallel installation of the 'Forge (dev)' version on the same phone
  as the release version.